requirements.txt
# Core web framework
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Deep Learning - TensorFlow/Keras (as requested)
tensorflow>=2.17.0
tf-keras>=2.17.0
numpy>=1.24.0
Pillow>=10.0.0

# HuggingFace Transformers for SegFormer
transformers>=4.35.0

# Additional utilities
requests>=2.28.0
python-dotenv>=1.0.0

# Azure Storage (for model files)
azure-storage-blob>=12.19.0

# Development
pytest>=7.0.0
black>=24.0.0